backend/requirements.txt
django>=4.2,<5
djangorestframework>=3.14
python-dotenv>=1.0
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
djoser>=2.2
pandas>=2.0
google-generativeai>=0.3
Pillow>=10.0
